refactor: start cleaning up conftest
